+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/CryptoPublicKey.kt
@@ -176,6 +176,9 @@ sealed class CryptoPublicKey : PemEncodable<Asn1Sequence>, Identifiable {
 
         val bits = n.bitLength().let { Size.of(it) ?: throw IllegalArgumentException("Unsupported key size $it bits") }
 
+        @Deprecated(message="Use a BigInteger-capable constructor instead", level = DeprecationLevel.ERROR)
+        constructor(n: ByteArray, e: Int): this(Asn1Integer.fromUnsignedByteArray(n), Asn1Integer(e) as Asn1Integer.Positive)
+
         constructor(n: Asn1Integer, e: Asn1Integer): this(n as Asn1Integer.Positive, e as Asn1Integer.Positive)
         constructor(n: BigInteger, e: BigInteger): this(n.toAsn1Integer(), e.toAsn1Integer())
         constructor(n: BigInteger, e: UInt): this(n.toAsn1Integer(), Asn1Integer(e))
@@ -340,6 +343,24 @@ sealed class CryptoPublicKey : PemEncodable<Asn1Sequence>, Identifiable {
             @Suppress("NOTHING_TO_INLINE")
             inline fun fromUncompressed(curve: ECCurve, x: ByteArray, y: ByteArray) =
                 ECPoint.fromUncompressed(curve, x, y).asPublicKey(false)
+
+            @Deprecated(
+                "Explicitly specify what you want",
+                ReplaceWith("fromCompressed(curve, x, usePositiveY)"),
+                DeprecationLevel.ERROR
+            )
+            @Suppress("NOTHING_TO_INLINE")
+            inline operator fun invoke(curve: ECCurve, x: ByteArray, usePositiveY: Boolean) =
+                fromCompressed(curve, x, usePositiveY)
+
+            @Deprecated(
+                "Explicitly specify what you want",
+                ReplaceWith("fromUncompressed(curve, x, y)"),
+                DeprecationLevel.ERROR
+            )
+            @Suppress("NOTHING_TO_INLINE")
+            inline operator fun invoke(curve: ECCurve, x: ByteArray, y: ByteArray) =
+                fromUncompressed(curve, x, y)
 
             /** Decodes a key from its ANSI X9.63 representation */
             @Throws(Throwable::class)