+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/CryptoSignature.kt
@@ -210,7 +210,12 @@ sealed interface CryptoSignature : Asn1Encodable<Asn1Element> {
                 return fromRS(r, s)
             }
 
-
+            @Deprecated(
+                "use fromRawBytes",
+                ReplaceWith("CryptoSignature.EC.fromRawBytes(input)"),
+                DeprecationLevel.ERROR
+            )
+            operator fun invoke(input: ByteArray): DefiniteLength = fromRawBytes(input)
 
         }