+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Rename `ObjectIdentifier.parse` -> `ObjectIdentifier.decodeFromAsn1ContentBytes` in accordance with other similar functions
 * Add dedicated Android targets (SDK 30 / JDK 1.8) to all modules
 * Fix internal deprecations
-* Remove deprecated functions:
+* Raise deprecation level to ERROR for deprecated functions:
     * `Asn1Element.Companion.parseAll`
     * `Asn1Element.Companion.parse`
     * `Asn1Element.Companion.decodeFromDerHexString`